requirements.txt
asgiref==3.8.1
asttokens==3.0.0
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.1.4
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-model-utils==5.0.0
django-polymorphic==3.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
docopt==0.6.2
executing==2.1.0
graphviz==0.20.3
idna==3.10
ipython==8.30.0
jedi==0.19.2
Markdown==3.7
matplotlib-inline==0.1.7
mysqlclient==2.2.6
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.10
